refactor(api): move UncancellableOrganizationInvitation error
in src/team/domain/error.js
