+++ api/lib/application/error-manager.js
@@ -287,10 +287,6 @@ function _mapToHttpError(error) {
     return new HttpErrors.ConflictError(error.message);
   }
 
-  if (error instanceof DomainErrors.UncancellableOrganizationInvitationError) {
-    return new HttpErrors.UnprocessableEntityError(error.message);
-  }
-
   if (error instanceof DomainErrors.UserShouldNotBeReconciledOnAnotherAccountError) {
     return new HttpErrors.UnprocessableEntityError(error.message, error.code, error.meta);
   }
