+++ api/src/team/domain/errors.js
@@ -1,6 +1,6 @@
-import { DomainError } from '../../shared/domain/errors.js';
+import { DomainError } from '../../../lib/domain/errors.js';
 
-export class UncancellableCertificationCenterInvitationError extends DomainError {
+class UncancellableCertificationCenterInvitationError extends DomainError {
   constructor(
     message = "L'invitation à ce centre de certification ne peut pas être annulée.",
     code = 'UNCANCELLABLE_CERTIFICATION_CENTER_INVITATION_CODE',
@@ -9,4 +9,19 @@ export class UncancellableCertificationCenterInvitationError extends DomainError
   }
 }
 
-export class MembershipNotFound extends DomainError {}
+class UncancellableOrganizationInvitationError extends DomainError {
+  constructor(
+    message = "L'invitation à cette organisation ne peut pas être annulée.",
+    code = 'UNCANCELLABLE_ORGANIZATION_INVITATION_CODE',
+  ) {
+    super(message, code);
+  }
+}
+
+class MembershipNotFound extends DomainError {}
+
+export {
+  MembershipNotFound,
+  UncancellableCertificationCenterInvitationError,
+  UncancellableOrganizationInvitationError,
+};
