+++ api/tests/unit/application/error-manager_test.js
@@ -23,7 +23,6 @@ import {
   OrganizationLearnerCannotBeDissociatedError,
   SendingEmailToInvalidDomainError,
   SendingEmailToInvalidEmailAddressError,
-  UncancellableOrganizationInvitationError,
   UnexpectedUserAccountError,
   UserHasAlreadyLeftSCO,
   UserShouldNotBeReconciledOnAnotherAccountError,
@@ -216,19 +215,6 @@ describe('Unit | Application | ErrorManager', function () {
 
       // then
       expect(HttpErrors.ForbiddenError).to.have.been.calledWithExactly(error.message, error.code);
-    });
-
-    it('should instantiate UnprocessableEntityError when UncancellableOrganizationInvitationError', async function () {
-      // given
-      const error = new UncancellableOrganizationInvitationError();
-      sinon.stub(HttpErrors, 'UnprocessableEntityError');
-      const params = { request: {}, h: hFake, error };
-
-      // when
-      await handle(params.request, params.h, params.error);
-
-      // then
-      expect(HttpErrors.UnprocessableEntityError).to.have.been.calledWithExactly(error.message);
     });
 
     it('should instantiate UnprocessableEntityError when UserShouldNotBeReconciledOnAnotherAccountError', async function () {
